+++ projects/demo/src/app/documentation/documentation.component.ts
@@ -1,8 +1,6 @@
-import { ChangeDetectionStrategy, Component } from '@angular/core';
+import { Component } from '@angular/core';
 
 @Component({
-  selector: 'app-documentation',
-  changeDetection: ChangeDetectionStrategy.OnPush,
   templateUrl: './documentation.component.html',
   styleUrl: './documentation.component.scss'
 })