+++ projects/demo/src/app/table-preview-native/table-preview-native.component.ts
@@ -5,10 +5,10 @@ import { MatButtonModule } from '@angular/material/button';
 import { MatSnackBarModule } from '@angular/material/snack-bar';
 import { RowActionsDirective } from '@softwarity/row-actions';
 import { BadgeComponent } from '../badge/badge.component';
-import { TablePreviewBase } from './table-preview-base';
+import { TablePreviewBase } from '../table-preview/table-preview-base';
 
 @Component({
-  selector: 'app-table-preview',
+  selector: 'app-table-preview-native',
   imports: [
     MatTableModule,
     MatIconModule,
@@ -17,11 +17,11 @@ import { TablePreviewBase } from './table-preview-base';
     RowActionsDirective,
     BadgeComponent,
   ],
-  templateUrl: './table-preview.component.html',
+  templateUrl: './table-preview-native.component.html',
   styles: [`
   .mat-column-name {
     border-left: 1px dashed var(--mat-sys-outline-variant);
   }
   `],
 })
-export class TablePreviewComponent extends TablePreviewBase {}
+export class TablePreviewNativeComponent extends TablePreviewBase {}